Changing Reddwarf client to use its own request function.
* Nova client changed in a way that broke our client, so copying the code from there is necessary.
* Adding InstanceStatus class with the status strings.
* Moved the Dbaas and ReddwarfHTTPClient into their own module.
* Changed exceptions module to check Nova's exception map after first looking in Reddwarf's.
